depth

+++ 10-binary_tree_depth.c
@@ -11,7 +11,7 @@ size_t binary_tree_depth(const binary_tree_t *tree)
 	if (!tree)
 		return (0);
 
-	int depth = 0;
+	size_t depth = 0;
 
 	if (tree->parent)
 		depth = binary_tree_depth(tree->parent) + 1;
